requirements.txt
flask
pillow
rembg
gunicorn